refactor, cleanup and add more doc comments
