+++ src/vs/editor/common/codecs/frontMatterCodec/tokens/index.ts
@@ -3,7 +3,8 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-export { FrontMatterToken } from './frontMatterToken.js';
 export { FrontMatterArray } from './frontMatterArray.js';
 export { FrontMatterString } from './frontMatterString.js';
 export { FrontMatterBoolean } from './frontMatterBoolean.js';
+export { FrontMatterToken, FrontMatterValueToken } from './frontMatterToken.js';
+export { FrontMatterRecordName, FrontMatterRecordDelimiter, FrontMatterRecord, type TNameToken as TRecordNameToken, type TSpaceToken as TRecordSpaceToken } from './frontMatterRecord.js';
