+++ src/vs/editor/common/codecs/markdownExtensionsCodec/parsers/frontMatterHeader.ts
@@ -5,11 +5,11 @@
 
 import { Dash } from '../../simpleCodec/tokens/dash.js';
 import { NewLine } from '../../linesCodec/tokens/newLine.js';
+import { FrontMatterHeader } from '../tokens/frontMatterHeader.js';
 import { assertDefined } from '../../../../../base/common/types.js';
 import { TSimpleDecoderToken } from '../../simpleCodec/simpleDecoder.js';
 import { assert, assertNever } from '../../../../../base/common/assert.js';
 import { CarriageReturn } from '../../linesCodec/tokens/carriageReturn.js';
-import { FrontMatterHeader } from '../tokens/frontMatterHeader.js';
 import { FrontMatterMarker, TMarkerToken } from '../tokens/frontMatterMarker.js';
 import { assertNotConsumed, IAcceptTokenSuccess, ParserBase, TAcceptTokenResult } from '../../simpleCodec/parserBase.js';
