+++ src/vs/editor/test/common/codecs/simpleDecoder.test.ts
@@ -34,6 +34,7 @@ import {
 	RightParenthesis,
 	LeftAngleBracket,
 	RightAngleBracket,
+	Comma,
 } from '../../../common/codecs/simpleCodec/tokens/index.js';
 
 /**
@@ -80,7 +81,7 @@ suite('SimpleDecoder', () => {
 				' hello world!',
 				'how are\t "you?"\v',
 				'',
-				'   (test)  [!@#$:%^🦄&*_+=-]\f  ',
+				'   (test)  [!@#$:%^🦄&*_+=,-,]\f  ',
 				'\t<hi 👋>\t🤗❤ \t',
 				' hey\v-\tthere\r',
 				' @workspace@legomushroom',
@@ -123,12 +124,14 @@ suite('SimpleDecoder', () => {
 				new DollarSign(new Range(4, 16, 4, 17)),
 				new Colon(new Range(4, 17, 4, 18)),
 				new Word(new Range(4, 18, 4, 18 + 9), '%^🦄&*_+='),
-				new Dash(new Range(4, 27, 4, 28)),
-				new RightBracket(new Range(4, 28, 4, 29)),
-				new FormFeed(new Range(4, 29, 4, 30)),
-				new Space(new Range(4, 30, 4, 31)),
-				new Space(new Range(4, 31, 4, 32)),
-				new NewLine(new Range(4, 32, 4, 33)),
+				new Comma(new Range(4, 27, 4, 28)),
+				new Dash(new Range(4, 28, 4, 29)),
+				new Comma(new Range(4, 29, 4, 30)),
+				new RightBracket(new Range(4, 30, 4, 31)),
+				new FormFeed(new Range(4, 31, 4, 32)),
+				new Space(new Range(4, 32, 4, 33)),
+				new Space(new Range(4, 33, 4, 34)),
+				new NewLine(new Range(4, 34, 4, 35)),
 				// fifth line
 				new Tab(new Range(5, 1, 5, 2)),
 				new LeftAngleBracket(new Range(5, 2, 5, 3)),
